Add optional fields
Add the ability to specify optional fields using the syntax

    name: type if expression,

with the semantics that if the expression evaluates to `true` the field
will be present, and if it evaluates to `false` it will not be. The
sizing and alignment requirements are the same as a variable-sized
array. In other words this can be viewed as syntactic sugar for

    name: [type; expression as u8],

This commit also fixes WrapErr::NotEnoughBytes to return a more-correct
number of bytes expected, and fixes expressions for variable-sized
arrays to work in more cases.